requirements.txt
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
confluent-kafka==2.3.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
ipython==8.21.0
jedi==0.19.1
kafka-python==2.0.2
matplotlib-inline==0.1.6
parso==0.8.3
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
requests==2.31.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
urllib3==2.2.1
wcwidth==0.2.13
